requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
boto3==1.35.90
certifi==2025.8.3
charset-normalizer==3.4.3
defusedxml==0.7.1
Django==5.2.5
django-crontab==0.7.1
django-rq==2.10.2
django-storages==1.14.4
django-tinymce==4.1.0
et_xmlfile==2.0.0
fonttools==4.60.1
fpdf2==2.7.9
frozenlist==1.7.0
gunicorn==22.0.0
idna==3.10
multidict==6.6.4
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
propcache==0.3.2
PyJWT==2.10.1
python-dotenv==1.1.1
qrcode==8.2
redis==5.0.1
requests==2.32.5
sqlparse==0.5.3
twilio==9.7.2
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
rq==1.15.1
